+++ NEWS.md
@@ -7,6 +7,7 @@
 * readDictionary(): add arguments "fileEncoding", "...", pass on to
   readLinesWithEncoding()
 * allow to set the cache directory with set_cache_dir()
+* preparePdf(): set paper to "special" if width or height are set
 
 # [kwb.utils 0.7.0](https://github.com/KWB-R/kwb.utils/releases/tag/v0.7.0) <small>2020-07-31</small>